backend/api/requirements.txt
fastapi[standard]
psycopg[binary]
passlib[bcrypt]
pyjwt
google-genai
